backend_api/requirement.txt
django
djangorestframework
djangorestframework_simplejwt
pillow
requests
psycopg2
sqlparse
dotenv